Refactor codebase into nmrkan package
