+++ requirements.txt
@@ -5,4 +5,4 @@ sympy>=1.11
 optuna>=3.0
 tqdm>=4.64
 matplotlib>=3.6
-seaborn>=0.12
+scipy>=1.9
