+++ run_symbolic_regressor.py
@@ -11,8 +11,8 @@ import sympy as sp
 import optuna
 import tqdm
 
-from src.kharkan.modelKAN import KharKAN, _clean_expr
-from src.kharkan.NMR import get_frequences, get_frequences_ordered
+from nmrkan.models import KharKAN, _clean_expr
+from nmrkan.nmr import get_frequences, get_frequences_ordered
 
 # --- Set up logging -----------------------------------------------
 logging.basicConfig(
